Fix flicker in most animated GIF
Hi Paul,

I set up an animated GIF panel in my son's room based on this demo of yours and initially I saw a lot of flicker in most GIF animations.

Eventually I found out that the `swapBuffers` needs to be called with the option to copy the current buffer into the next buffer after doing the swap, otherwise the frame get out of sync with the GIF decoder causing flicker.

Did you get this code from somewhere else or did you write it? It's really nice. It would be worth it to package this up as a demo app for the RGB shield.